refactor: pagination common 으로 빼기
